docs(store): add `createFeature` docs (#3057)
* docs(store): add createFeature docs

* fix document generation for createFeature
